Struct opencl::hl::Program

pub struct Program {
    // some fields omitted
}

Represents an OpenCL program, which is a collection of kernels.

Create these using Context::create_program_from_source or Context::create_program_from_binary.

Methods

impl Program

fn build(&self, device: &Device) -> Result<String, String>

Build the program for a given device with given compiler options.

Both Ok and Err returns include the build log.

fn build_with_options(&self, device: &Device, options: &str) -> Result<String, String>

Build the program for a given device.

Both Ok and Err returns include the build log.

fn create_kernel(&self, name: &str) -> Kernel